refactor(mobile): extract LiveRoundSession; callback-driven hole nav (commit A of #264)
First step toward making the live-round MapView resident across hole
transitions. Today this commit is behavior-preserving — the route is
still path-segmented, the hole screen still remounts on Prev/Next.
The structure is now in place to mount LiveRoundSession from a
non-remounting location in a later commit.

- New components/round/LiveRoundSession.tsx — owns holeNumber as
  internal useState, hosts the full hole-screen body that previously
  lived in app/(app)/round/[id]/hole/[number].tsx. Per-hole modal
  flags reset via useEffect([holeNumber]); the hook-owned refs
  (Kalman state, manualPlaced, lastSavedShotLocalId) already reset on
  currentHoleId change inside useHoleState — no duplication needed.

- useShotActions now takes onHoleChange: (next) => void. navigateHole
  and finishHole call the callback instead of router.replace; hole 18
  finishHole still router.replace's out to /(app). Today's caller is
  the route wrapper passing a router-replace closure so behavior is
  preserved.

- useHoleCamera gets an explicit "new hole detected" flyTo effect so
  the resident camera animates to the new tee instead of relying on
  the per-hole component remount it used to assume. The existing
  initial-frame and per-phase setCamera sites are untouched.

- The old [number].tsx route file is now a 25-line wrapper that
  reads route params and mounts LiveRoundSession with a router-replace
  onHoleChange.

Commits B and C will retarget internal URL builders to /round/[id]?hole=N
and then move the LiveRoundSession mount to the index route, deleting
the path-segmented hole route.
